[#58] Refactor and moved indicator state to global
