+++ frontend/src/state/ui.js
@@ -4,6 +4,8 @@ const defaultUIState = {
   advanceSearchValue: [],
   provinceValues: [],
   schoolTypeValues: [],
+  indicatorQuestions: [],
+  advanceFilterQuestions: [],
 };
 
 export const UIState = new Store(defaultUIState);
